Move findProfanity(...) to common utils file
