+++ apps/src/utils.js
@@ -817,6 +817,22 @@ export function escapeRegExp(str) {
 }
 
 /**
+ * Detects profanity in a string in a given language.
+ * @param {string} text
+ * @param {string} language
+ * @param {bool} shouldCache Whether or not the server should cache the response. Optional.
+ * @returns {Array<string>|null} Array of profane words.
+ */
+export const findProfanity = (text, language, shouldCache) => {
+  return $.ajax({
+    url: '/profanity/find',
+    method: 'POST',
+    contentType: 'application/json;charset=UTF-8',
+    data: JSON.stringify({text, language, should_cache: shouldCache})
+  });
+};
+
+/**
  * Convert a string to an MD5 hash.
  * @param {string} str
  * @returns {string} A string representing an MD5 hash.
